requirements.txt
python-telegram-bot==20.7
supabase==2.3.4
python-dotenv==1.0.0
qrcode==7.4.2
Pillow==10.2.0
phonenumbers==8.13.27
validators==0.22.0
